refactor(checker): split heritage constructor/before-decl checks into sibling module
heritage.rs grew to 2008 lines after the re-export alias-chain fix, tripping
the 2000-line checker-boundary arch guard. Move the lexical-enclosure,
type-parameter-reference, before-declaration, and heritage call-expression
constructor-compatibility checks (plus their tests) into a new
heritage_constructor_checks.rs sibling module. Behavior unchanged; the five
cross-called methods are bumped to pub(crate).
